requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

google_auth_oauthlib
googleapiclient
google-cloud-storage
requests
beautifulsoup4
html5lib
joblib

